add token for delete Pan

+++ src/Controller/PansController.php
@@ -85,8 +85,11 @@ class PansController extends AbstractController
      /**
      * @Route("/pans/delete/{id<[0-9+]>}", name="app_pans_delete", methods={"DELETE"})
      */
-    public function delete(Pan $pan, EntityManagerInterface $em): Response
+    public function delete(Request $request, Pan $pan, EntityManagerInterface $em): Response
     {
+        $token = $request->request->get('csrf_token');
+        dd($token);
+        if($this->isCsrfTokenValid('pan_delete_' . $pan->getId(), $token))
         $em->remove($pan);
         $em->flush();
         return $this->redirectToRoute('app_home');
